Don't apply update when update already in progress
Fixes #8

Some events in nerves-hub.org will trigger an update message to be sent, such as editing
a deployment/device (changing the name and saving, etc).

If a device gets that message while it is already updating, things will crash. This shouldn't
be too much of an issue because the whole channel crashes, reconnects, and gets the update
message again. But it produces some nasty error that might be confusing to the user.

So to help avoid this edge-case, let's just ignore update messages while in the process of
updating. Should it actually be new/required, we'll get the message again on next startup after
rebooting.