Fix timeout when reconnecting
This removes an unnecessary synchronous call to await_disconnect that
would time out when there wasn't a connection to disconnect. The easier
way to handle this is to not bother waiting at all and let the
reconnection logic work.

This looks funny so I left a note. This fixes the exception that was
being thrown when timing out and it reconnects properly when reconnect
is called both when there's a connection and not.