Parameterize the exponential backoff
The list of delays passed to Slipstream is now parameterized by a
minimum delay, a maximum and a jitter. The default is to first delay for
a second and then progressively double it until a 60 second delay.

Jitter is specified as a percentage to add. For 0.5 jitter, this means
that what would normally be a 1 second delay might be up to 1.5 seconds.
Similarly, a 10 second delay might be as long as 15 seconds and so on.

The delays are seeded from `:crypto.strong_rand_bytes/1` to help ensure
that there's a good spread across devices. While `:rand.seed/2`'s
defaults may be ok, there was concern since it's seeded by the system
time.