Use `Slipstream.push/5` for safe publishing
We've recently run into a condition where another failure caused the socket to
crash, but `fwup` was still applying an update sending messages to new socket
process which hadn't joined the channel yet. `push!/4` would then crash
the socket again and repeat the cycle.

In most cases, these messages are not critical so switch to the safe `push/5`
version, we can avoid needless crashes and still be okay on the web side