Remove `NervesHubLink.Connection`
This was previously an Agent for minimal connection state that was
used due to the limitations of `phoenix_client` and socket tracking.
However, since the switch to `slipstream` is has been unneeded and
virtually unused.

There was also some behavior of checking "uptime" of the connection
which is removed with this change in favor of NHL users handling
that check with their own logic. Times for process start, connection, and successful join are now saved in the socket instead of the Agent
for future use