ArchiveManager to handle NervesHub sending extra firmware (#135)
* ArchiveManager to handle NervesHub sending extra firmware

Archives are an extra firmware attached to a deployment. On connect
NervesHub will potentially send a message about an archive and the
device can download the firmware to a configured / known location and
tell the configured client that it's there. ArchiveManager does nothing
else as the extra firmware will be highly application specific.

* fixup! ArchiveManager to handle NervesHub sending extra firmware