Add extension mechanism with process isolation from main Socket
The goal is to create a mechanism allowing us to re-use the existing connection
and add new functionality based on messages received and passing messages back
up the WebSocket.

The important part is that new features and functionality that are not core should
be possible to deliver as libraries and plugged in via configuration. They should
not be able to crash the Socket and prevent firmware updates from succeeding.

This is not complete isolation. Things could still be brought down by a really broken
extension that brings down the :nerves_hub_link app but extensions run quite well
separated.