Jon late night, half ideas
This starts the idea of using an extension(s) channel and a behavior
for routing messages to and from extensions. This could allow the
extension to be a simple module or a full baked application of
genservers to handle complex situations.

This could also potentially support one or more extension processes
on the server with setups like

```elixir
channel("extension:my_ext", MyApp.ExtChannel)
channel("extension:*", NervesHubWeb.ExtensionChannel)
```