Change to only configure if connecting
Currently the configurator would be built even if there was no
connection planned which meant the :nerves_hub_link application
could easily die when it was supposed to do nothing.

This change will not process the config with a configurator when
connect has been set to false.

It enables dynamic onboarding to NervesHub.