Fix the config key used for fetching the `retry_config`
And improve the docs for the `Downloader` and `RetryConfig` modules.

The retry logic was merged in from a lib called `nerves_hub_link_common` many moons ago, but the key used for fetching the config and the version from the spec were not updated.

This fixes the `User-Agent` header, and aligns how the retry logic should be configured.