Revert Backoff/Certificate move to fix breaking change
These really looked like private API due to the `@moduledoc false`.
However, they're actually in use in production software. I was going to
fix the calls in the firmware, but I'd be copy/pasting all of the code
in these files. Therefore, I'm moving back and adding a short
`@moduledoc` to make it more obvious they're public API.